More bucket computation via linear scan
In addition to scanning the PGM Index, we can also scan the buckets:
We do not need to distribute all the elements to the buckets
before starting to construct buckets. If we come across one element
with a larger bucket number, we can already construct the previous
bucket. That saves us a lot of peak memory because we only need to
keep a single bucket active at all times.

Makes the construction about 10% faster.