Add some default alphabet maps
This is based on the idea that if we have an alphabet map
with 130 characters, we might as well just use one that contains
all 256 characters. The number of characters we can squeeze into a
chunk is the same. Now, the alphabet reuse code might pick these
default maps first.

The prepared "letters" alphabet maps currently don't make a difference,
but I figured they don't hurt either. My main goal was to have the full
alphabet maps in there, which enables future optimizations.